+++ client/cody-jetbrains/src/main/java/com/sourcegraph/cody/api/CompletionsCallbacks.java
@@ -1,0 +1,14 @@
+package com.sourcegraph.cody.api;
+
+// Define a callback interface to handle events
+public interface CompletionsCallbacks {
+  void onSubscribed();
+
+  void onData(String data);
+
+  void onError(Throwable error);
+
+  void onComplete();
+
+  void onCancelled();
+}