+++ client/cody-ui/src/chat/CodeBlocks.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useMemo } from 'react'
+import React, { useEffect, useMemo, useRef } from 'react'
 
 import classNames from 'classnames'
 
@@ -12,6 +12,7 @@ interface CodeBlocksProps {
     displayText: string
 
     copyButtonClassName?: string
+    insertButtonClassName?: string
 
     CopyButtonProps?: CopyButtonProps['copyButtonOnSubmit']
 }
@@ -24,55 +25,124 @@ function wrapElement(element: HTMLElement, wrapperElement: HTMLElement): void {
     wrapperElement.append(element)
 }
 
-function createCopyButtonWithContainer(
+function createButtons(
     text: string,
-    className: string,
+    copyButtonClassName?: string,
+    copyButtonOnSubmit?: CopyButtonProps['copyButtonOnSubmit'],
+    insertButtonClassName?: string
+): HTMLElement {
+    const container = document.createElement('div')
+    container.className = styles.container
+
+    // The container will contain the buttons and the <pre> element with the code.
+    // This allows us to position the buttons independent of the code.
+    const buttons = document.createElement('div')
+    buttons.className = styles.buttons
+
+    const copyButton = createCopyButton(text, copyButtonClassName, copyButtonOnSubmit)
+    const insertButton = createInsertButton(text, container, insertButtonClassName, copyButtonOnSubmit)
+
+    // The insert button only exists for IDE integrations
+    if (insertButton) {
+        buttons.append(insertButton)
+    }
+    buttons.append(copyButton)
+
+    container.append(buttons)
+
+    return container
+}
+
+function createCopyButton(
+    text: string,
+    className?: string,
     copyButtonOnSubmit?: CopyButtonProps['copyButtonOnSubmit']
 ): HTMLElement {
-    const copyButton = document.createElement('button')
-    copyButton.textContent = 'Copy'
-    copyButton.className = className
-    copyButton.addEventListener('click', () => {
+    const button = document.createElement('button')
+    button.textContent = 'Copy'
+    button.title = 'Copy text'
+    button.className = classNames(styles.copyButton, className)
+    button.addEventListener('click', () => {
         navigator.clipboard.writeText(text).catch(error => console.error(error))
-        copyButton.textContent = 'Copied!'
-        setTimeout(() => (copyButton.textContent = 'Copy'), 3000)
+        button.textContent = 'Copied'
+        setTimeout(() => (button.textContent = 'Copy'), 3000)
         if (copyButtonOnSubmit) {
             copyButtonOnSubmit('copyButton')
         }
     })
+    return button
+}
 
-    // The container will contain the copy button and the <pre> element with the code.
-    // This allows us to position the copy button independent of the code.
-    const container = document.createElement('div')
-    container.className = styles.container
-    container.append(copyButton)
-    return container
+function createInsertButton(
+    text: string,
+    container: HTMLElement,
+    className?: string,
+    copyButtonOnSubmit?: CopyButtonProps['copyButtonOnSubmit']
+): HTMLElement | null {
+    if (!className || !copyButtonOnSubmit) {
+        return null
+    }
+    const button = document.createElement('button')
+    button.textContent = 'Insert at Cursor'
+    button.title = 'Insert text at current cursor position'
+    button.className = classNames(styles.insertButton, className)
+    button.addEventListener('click', () => {
+        const selectedText = getSelectedTextWithin(container.querySelector('pre'))
+        copyButtonOnSubmit(selectedText || text, true)
+    })
+    return button
+}
+
+function getSelectedTextWithin(element: HTMLElement | null): string | null {
+    if (!element) {
+        return null
+    }
+
+    const selection = document.getSelection()
+    if (!selection) {
+        return null
+    }
+
+    const range = selection.getRangeAt(0)
+    const startContainer = range.startContainer
+    const endContainer = range.endContainer
+
+    if (element.contains(startContainer) && element.contains(endContainer)) {
+        return selection.toString()
+    }
+
+    return null
 }
 
 export const CodeBlocks: React.FunctionComponent<CodeBlocksProps> = ({
     displayText,
     copyButtonClassName,
+    insertButtonClassName,
     CopyButtonProps,
 }) => {
+    const rootRef = useRef<HTMLDivElement>(null)
+
     useEffect(() => {
-        const preElements = document.querySelectorAll('pre')
+        const preElements = rootRef.current?.querySelectorAll('pre')
+        if (!preElements?.length) {
+            return
+        }
+
         for (const preElement of preElements) {
             const preText = preElement.textContent
-            const hasCopyButton = preElement.querySelector(`.${styles.container}`)
-            if (!hasCopyButton && preText && preText.trim().length > 0) {
-                // We have to wrap the `<pre>` tag in the copy button container, otherwise
-                // the Copy button scrolls along with the code.
+            if (preText?.trim()) {
+                // We have to wrap the `<pre>` tag in the button container, otherwise
+                // the buttons scroll along with the code.
                 wrapElement(
                     preElement,
-                    createCopyButtonWithContainer(
-                        preText,
-                        classNames(styles.copyButton, copyButtonClassName),
-                        CopyButtonProps
-                    )
+                    createButtons(preText, copyButtonClassName, CopyButtonProps, insertButtonClassName)
                 )
             }
         }
-    }, [copyButtonClassName, displayText, CopyButtonProps])
+    }, [displayText, CopyButtonProps, copyButtonClassName, insertButtonClassName, rootRef])
 
-    return useMemo(() => <div dangerouslySetInnerHTML={{ __html: renderCodyMarkdown(displayText) }} />, [displayText])
+    return useMemo(
+        () => <div ref={rootRef} dangerouslySetInnerHTML={{ __html: renderCodyMarkdown(displayText) }} />,
+        [displayText]
+    )
 }