+++ client/cody/src/chat/ChatViewProvider.ts
@@ -24,6 +24,7 @@ import { VSCodeEditor } from '../editor/vscode-editor'
 import { logEvent } from '../event-logger'
 import { LocalAppDetector } from '../local-app-detector'
 import { debug } from '../log'
+import { FixupTask } from '../non-stop/FixupTask'
 import { LocalStorage } from '../services/LocalStorageProvider'
 import { CODY_ACCESS_TOKEN_SECRET, SecretStorage } from '../services/SecretStorageProvider'
 import { TestSupport } from '../test-support'
@@ -239,6 +240,9 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
                 void this.webview?.postMessage({ type: 'login', authStatus })
                 break
             }
+            case 'insert':
+                await vscode.commands.executeCommand('cody.inline.insert', message.text)
+                break
             case 'event':
                 this.sendEvent(message.event, message.value)
                 break
@@ -305,9 +309,10 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
                     // TODO(keegancsmith) guardrails may be slow, we need to make this async update the interaction.
                     highlightedDisplayText = await this.guardrailsAnnotateAttributions(highlightedDisplayText)
                     this.transcript.addAssistantResponse(text || '', highlightedDisplayText)
-                    this.editor.controller.reply(highlightedDisplayText)
+                    this.editor.controllers.inline.reply(highlightedDisplayText)
                 }
                 void this.onCompletionEnd()
+                this.publishEmbeddingsError()
             },
         })
 
@@ -362,6 +367,13 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
         this.sendTranscript()
         void this.saveTranscriptToChatHistory()
         void vscode.commands.executeCommand('setContext', 'cody.reply.pending', false)
+        if (!this.codebaseContext.checkEmbeddingsConnection()) {
+            this.publishEmbeddingsError()
+            this.sendErrorToWebview(
+                'Error while establishing embeddings server connection. Please try after sometime! If the issue still persists contact support'
+            )
+            return
+        }
     }
 
     private async onHumanMessageSubmitted(text: string, submitType: 'user' | 'suggestion'): Promise<void> {
@@ -650,9 +662,19 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
                 },
             })
         }
-
         this.disposables.push(vscode.window.onDidChangeTextEditorSelection(() => send()))
         send()
+    }
+
+    /**
+     * Publish embedding connections or results error to webview
+     */
+    private publishEmbeddingsError(): void {
+        const searchErrors = this.codebaseContext.getEmbeddingSearchErrors()
+        if (searchErrors.length) {
+            this.sendErrorToWebview(searchErrors)
+            return
+        }
     }
 
     /**
@@ -791,6 +813,14 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
             return []
         }
         return this.transcript.toChat()
+    }
+
+    public fixupTasksForTesting(testing: TestSupport): FixupTask[] {
+        if (!testing) {
+            console.error('used ForTesting method without test support object')
+            return []
+        }
+        return this.editor.controllers.task.getTasks()
     }
 
     public dispose(): void {